+++ next-docs/package.json
@@ -24,6 +24,7 @@
     "@heathmont/moon-datepicker": "workspace:^9.30.1",
     "@heathmont/moon-draggabletable": "workspace:^9.30.1",
     "@heathmont/moon-icons": "workspace:^9.30.1",
+    "@heathmont/moon-icons-tw": "9.28.3",
     "@heathmont/moon-pagination": "workspace:^9.30.1",
     "@heathmont/moon-popover": "workspace:^9.30.1",
     "@heathmont/moon-select": "workspace:^9.30.1",
@@ -33,7 +34,6 @@
     "@heathmont/moon-themes": "workspace:^9.30.1",
     "@heathmont/moon-themes-tw": "workspace:^9.30.1",
     "@heathmont/moon-utils": "workspace:^9.30.1",
-    "@heathmont/moon-icons-tw": "9.28.3",
     "@jest/globals": "27.4.6",
     "@next/bundle-analyzer": "^13.0.1",
     "@octokit/core": "3.5.1",
@@ -58,14 +58,14 @@
     "react": "18.2.0",
     "react-copy-to-clipboard": "5.0.4",
     "react-dom": "18.2.0",
+    "react-syntax-highlighter": "15.5.0",
     "react-test-renderer": "18.0.0",
     "sharp": "0.30.6",
     "styled-components": "5.3.5",
     "swr": "1.0.1",
     "tailwindcss": "3.2.3",
     "tailwindcss-rtl": "0.9.0",
-    "typescript": "4.9.5",
-    "react-syntax-highlighter": "15.5.0"
+    "typescript": "4.9.5"
   },
   "devDependencies": {
     "@svgr/core": "6.2.1",
@@ -73,6 +73,7 @@
     "@svgr/plugin-prettier": "6.1.2",
     "@types/lodash": "^4.14.182",
     "@types/react-syntax-highlighter": "15.5.6",
+    "duplicate-package-checker-webpack-plugin": "^3.0.0",
     "snapshot-diff": "0.9.0"
   }
 }
