+++ next-docs/tsconfig.json
@@ -1,10 +1,6 @@
 {
   "compilerOptions": {
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -20,17 +16,8 @@
     "incremental": true
   },
   "paths": {
-    "@heathmont/moon-*": [
-      "./packages/*/lib",
-      "./next-docs"
-    ]
+    "@heathmont/moon-*": ["./packages/*/lib", "./next-docs"]
   },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "utils/analyzer.js"],
+  "exclude": ["node_modules"]
 }
